fix mock fraud sdk
